fix guid

+++ src/PulseEngine/core/FileManager/Archive/Archive.h
@@ -18,7 +18,7 @@ public:
 
     // Core API for primitive types
     virtual void Serialize(const char* name, int& value) = 0;
-    virtual void Serialize(const char* name, std::size_t& value) = 0;
+    virtual void Serialize(const char* name, std::uint64_t& value) = 0;
     virtual void Serialize(const char* name, float& value) = 0;
     virtual void Serialize(const char* name, uint32_t& value) = 0;
     virtual void Serialize(const char* name, std::string& value) = 0;
